feat: add exception handler to auth service

+++ auth-service/src/main/java/com/healthcare/auth_service/service/AuthServiceImpl.java
@@ -4,32 +4,39 @@ import com.healthcare.auth_service.domain.AuthUserDetails;
 import com.healthcare.auth_service.domain.dto.LoginDto;
 import com.healthcare.auth_service.domain.dto.RegistrationDto;
 import com.healthcare.auth_service.domain.dto.TokensDto;
+import com.healthcare.auth_service.exception_handler.exception.AccessDeniedException;
+import com.healthcare.auth_service.exception_handler.exception.UnauthorizedException;
 import com.healthcare.auth_service.service.feignClient.UserClient;
 import com.healthcare.auth_service.service.interfacies.AuthService;
 import com.healthcare.auth_service.service.interfacies.BlockService;
 import com.healthcare.auth_service.service.interfacies.RefreshTokenService;
 import com.healthcare.auth_service.service.interfacies.TokenBlacklistService;
 import jakarta.servlet.http.HttpServletRequest;
+import lombok.AccessLevel;
 import lombok.RequiredArgsConstructor;
+import lombok.experimental.FieldDefaults;
 import org.springframework.security.authentication.AuthenticationManager;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
+import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
+import org.springframework.util.StringUtils;
 
 import static com.healthcare.auth_service.service.token_utilities.TokenUtilities.extractJwtFromRequest;
 
 @Service
 @RequiredArgsConstructor
+@FieldDefaults(level = AccessLevel.PRIVATE, makeFinal = true)
 public class AuthServiceImpl implements AuthService {
 
-    private final UserClient userClient;
-    private final PasswordEncoder passwordEncoder;
-    private final JwtService jwtService;
-    private final AuthenticationManager authManager;
-    private final RefreshTokenService refreshTokenService;
-    private final CookieService cookieService;
-    private final BlockService blockService;
-    private final TokenBlacklistService tokenBlacklistService;
+    UserClient userClient;
+    PasswordEncoder passwordEncoder;
+    JwtService jwtService;
+    AuthenticationManager authManager;
+    RefreshTokenService refreshTokenService;
+    CookieService cookieService;
+    BlockService blockService;
+    TokenBlacklistService tokenBlacklistService;
 
     @Override
     public TokensDto registeration(RegistrationDto dto) {
@@ -44,16 +51,17 @@ public class AuthServiceImpl implements AuthService {
     @Override
     public TokensDto login(LoginDto dto) {
 
-        var auth = new UsernamePasswordAuthenticationToken(dto.getUserEmail(), dto.getPassword());
-        authManager.authenticate(auth);
-
         AuthUserDetails userDetails = userClient.getUserByEmail(dto.getUserEmail());
 
         if (blockService.isBlocked(userDetails.getId())) {
-            throw new RuntimeException("Превышен лимит активных сессий. Попробуйте позже.");
+            throw new AccessDeniedException("The limit of active sessions is exceeded. Try it later.");
         }
 
+        var auth = new UsernamePasswordAuthenticationToken(dto.getUserEmail(), dto.getPassword());
+        authManager.authenticate(auth);
+
         return generateAndStoreTokens(userDetails);
+
     }
 
     @Override
@@ -63,23 +71,28 @@ public class AuthServiceImpl implements AuthService {
 
         AuthUserDetails userDetails = validateRefreshTokenAndGetUser(refreshToken);
 
-        TokensDto tokens = generateAndStoreTokens(userDetails);
-
         refreshTokenService.delete(refreshToken, userDetails.getId());
 
-        return tokens;
+        return generateAndStoreTokens(userDetails);
     }
 
     @Override
     public void logout(HttpServletRequest request) {
 
-        tokenBlacklistService.blacklist(extractJwtFromRequest(request));
+        String accessToken = extractJwtFromRequest(request);
+
+        if (StringUtils.hasText(accessToken)) {
+            long ttl = jwtService.getRemainingTTLAccessToken(accessToken);
+            tokenBlacklistService.blacklist(accessToken, ttl);
+        }
 
         String refreshToken = cookieService.getRefreshTokenFromCookie(request);
 
-        AuthUserDetails userDetails = validateRefreshTokenAndGetUser(refreshToken);
-
-        refreshTokenService.delete(refreshToken, userDetails.getId());
+        if (StringUtils.hasText(refreshToken)) {
+            AuthUserDetails userDetails = validateRefreshTokenAndGetUser(refreshToken);
+            refreshTokenService.delete(refreshToken, userDetails.getId());
+        }
+        SecurityContextHolder.clearContext();
     }
 
     private TokensDto generateAndStoreTokens(AuthUserDetails userDetails) {
@@ -92,20 +105,25 @@ public class AuthServiceImpl implements AuthService {
 
     private AuthUserDetails validateRefreshTokenAndGetUser(String refreshToken) {
 
-        if (refreshToken == null || refreshToken.isBlank()) {
-            throw new RuntimeException("No refresh token provided");
+        if (!StringUtils.hasText(refreshToken)) {
+            throw new UnauthorizedException("No refresh token provided");
         }
 
-        String email = jwtService.extractUserEmailFromRefreshToken(refreshToken);
+        String email;
+        try {
+            email = jwtService.extractUserEmailFromRefreshToken(refreshToken);
+        } catch (Exception e) {
+            throw new UnauthorizedException("Invalid refresh token");
+        }
 
         AuthUserDetails userDetails = userClient.getUserByEmail(email);
 
         if (!jwtService.validateRefreshToken(refreshToken, userDetails)) {
-            throw new RuntimeException("Invalid or expired token");
+            throw new UnauthorizedException("Invalid or expired token");
         }
 
         if (!refreshTokenService.isValid(refreshToken, userDetails.getId())) {
-            throw new RuntimeException("Refresh token not found");
+            throw new UnauthorizedException("Refresh token not found or revoked");
         }
         return userDetails;
     }
